data_collection/scrapers/generic_scraper.py
import requests
from bs4 import BeautifulSoup
import time
import random

from core.models import PricePoint
from data_collection.base import BaseCollector

class GenericScraper(BaseCollector):
    """Generic web scraper for exchanges without APIs."""
    
    def __init__(self, exchange_name, base_url, p2p_url=None):
        self.exchange_name = exchange_name
        self.base_url = base_url
        self.p2p_url = p2p_url or base_url
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36',
            'Accept-Language': 'en-US,en;q=0.9',
            'Accept-Encoding': 'gzip, deflate, br',
            'Connection': 'keep-alive'
        }
    
    def _make_request(self, url, params=None):
        """Make a request to the website with proper headers and random delay."""
        # Add a small random delay to avoid rate limiting
        time.sleep(random.uniform(1, 3))
        
        try:
            response = requests.get(url, headers=self.headers, params=params)
            response.raise_for_status()
            return response
        except Exception as e:
            print(f"Error making request to {url}: {e}")
            return None
    
    def fetch_p2p_prices(self, asset):
        """Scrape P2P prices from the website."""
        url = f"{self.p2p_url}/p2p/market/{asset.lower()}"
        response = self._make_request(url)
        
        if not response:
            return []
        
        soup = BeautifulSoup(response.text, 'html.parser')
        price_points = []
        
        # Example: find all P2P listings
        listings = soup.select('.p2p-listing')
        
        for listing in listings:
            try:
                # Extract data from HTML elements (customize based on the site structure)
                price_element = listing.select_one('.price')
                price = float(price_element.text.strip().replace('$', ''))
                
                # Create price point
                price_point = PricePoint(
                    exchange=self.exchange_name,
                    asset=asset,
                    price=price,
                    order_type="BUY",  # Would need to extract from the page
                    market_type="P2P",
                    available_amount=float(listing.select_one('.amount').text.strip().split()[0])
                )
                
                price_points.append(price_point)
            except Exception as e:
                print(f"Error parsing listing: {e}")
                continue
        
        return price_points
    
    # Implement other required methods with similar web scraping logic
    def fetch_exchange_prices(self, asset):
        """Scrape exchange prices from the website."""
        # Implementation details...
        pass
    
    def fetch_available_amount(self, asset, order_type):
        """Scrape available amount from the website."""
        # Implementation details...
        pass